disCovErr.jl Documentation

disCovErr.jl Documentation

disCovErr.add_sky_noise! — Method

add_sky_noise!(testim2,maskim0,skyim3,gain;seed=2021)

Adds noise to the infill that matches the Poisson noise of a rough estimate for the sky background. A random seed to set a local random generator is provided for reproducible unit testing.

Arguments:

testim2: input image which had infilling

maskim0: mask of pixels which were infilled

skyim3: rough estimate of sky background counts

gain: gain of detector to convert from photon count noise to detector noise

seed: random seed for random generator

disCovErr.boxsmoothMod! — Method

boxsmoothMod!(out, arr, widx::Int, widy::Int, sx::Int, sy::Int)

Boxcar smooths an input image (or paddedview) arr with window size widx by widy. We pass the original image size sx and sy to help handle image views.

Arguments:

out: preallocated output array for the boxcar smoothed image

arr: input array for which boxcar smoothing is computed (generally paddedview)

widx::Int: size of boxcar smoothing window in x

widy::Int: size of boxcar smoothing window in y

sx::Int: x size of the original (unpadded) image

sy::Int: y size of the original (unpadded) image

disCovErr.condCovEst_wdiag — Method

condCovEst_wdiag(cov_loc,μ,k,kstar,kpsf2d,data_in,data_w,stars_in) -> [std_w std_wdiag var_wdb resid_mean pred_mean chi20]

Using a local covariance matrix estimate cov_loc and a set of known pixels k and unknown pixels kstar, this function computes a prediction for the mean value of the kstar pixels and the covariance matrix of the kstar pixels. In terms of statistics use to adjust the photometry of a star, we are only interested in the pixels masked as a result of the star (i.e. not a detector defect or cosmic ray nearby) which is kpsf2d. The residual image data_in, the weight image data_w, and a model of the counts above the background coming from the star stars_in for the local patch are also inputs of the function. Correction factors for the photometric flux and flux uncertainities are outputs as well as the chi2 value for the predicted pixels.

Arguments:

cov_loc: local covariance matrix

μ: vector containing mean value for each pixel in the patch

k: unmasked pixels

kstar: masked pixels

kpsf2d: pixels masked due to the star of interest

data_in: residual image in local patch

data_w: weight image in local patch

stars_in: image of counts from star alone in local patch

disCovErr.cov_construct — Method

cov_construct(img, cx, cy; Np::Int=33, widx::Int=129, widy::Int=129) -> cov, μ

Construct a local covariance matrix estimate from img centered around pixel cx and cy. The covariance matrix will be for a square subimage of size Np by Np pixels, yielding cov of size Np^2. The covariance matrix is estimated by taking samples over pixels within a box of size widx and widy centered on the pixel of interest.

If cx and cy are arrays, the first index of the returned cov will be the index of the center point and μ will be the same length as cx.

In the interest of speed for the case of wanting to construct the covariance matrix at many posiitons, the moving mean of the full input image (times shifts of the input image) is precomputed for every entry of the covariance matrix. This may not be efficient for few cx.

Note that the values of widx and widy determine how local an estimate the covariance matrix returned is, but for too small of values of widx and widy, the covariance matrix estimate may be singular (one should always keep Np^2 < widx*widy).

Arguments:

img: Input image for which we desire the covariance matrix estimate

cx: x position (or positions) at which we want a covariance matrix estimate

cy: y position (or positions) at which we want a covariance matrix estimate

Np: size of the image stamp for which the covariance matrix will be estimated

widx: size of sampled region in x used to estimate entries in the covariance matrix

widy: size of sampled region in y used to estimate the entries in the covariance matrix

disCovErr.gen_mask_staticPSF! — Function

gen_mask_staticPSF!(maskd, psfstamp, x_stars, y_stars, flux_stars, thr=20)

Generate a mask for an input image (which is usually an image of model residuals) that excludes the cores of stars (which are often mismodeled). In this function, we use a fixed PSF psfstamp for all sources, and adjust the masking fraction based on the stellar flux and a threshold thr. A more general position dependent PSF model could be used with a slight generalization of this function, but is likely overkill for the problem of making a mask.

Arguments:

maskd: bool image to which mask will be added (bitwise or)

psfstamp: simple 2D array of a single PSF to be used for the whole image

x_stars: list of source x positions

y_stars: list of source y positions

flux_stars: list of source fluxes

thr: threshold used for flux-dependent masking

disCovErr.prelim_infill! — Method

prelim_infill!(testim,maskim,bimage,bimageI,testim2, maskim2, goodpix; widx = 19, widy=19)

This intial infill replaces masked pixels with a guess based on a smoothed boxcar. For large masked regions, the smoothing scale is increased. If this iteration takes too long/requires too strong of masking, the masked pixels are replaced with the median of the image.

We use 3 copies of the input image and mask image. The first is an untouched view (with reflective boundary condition padding), the second is allocated to hold various smoothings of the image, and the third holds the output image which contains our best infill guess. A final bool array of size corresponding to the image is used to keep track of pixels that have safe infill values.

Arguments:

testim: input image which requires infilling

bimage: preallocated array for smoothed version of input image

testim2: inplace modified ouptut array for infilled version of image input

maskim: input mask indicating which pixels require infilling

bimageI: preallocated array for smoothed mask counting the samples for each estimate

maskim2: inplace modified mask to keep track of which pixels still need infilling

widx::Int: size of boxcar smoothing window in x

widy::Int: size of boxcar smoothing window in y

disCovErr.read_crowdsource — Method

read_crowdsource(base,date,filt,vers,ccd) -> x_stars, y_stars, decapsid, gain, mod_im, sky_im

Read in outputs of crowdsource, a photometric pipeline. To pair with an arbitrary photometric pipeline, an analogous read in function should be created. The relevant outputs are the model image (including the sources) so that we can produce the residual image, the sky/background model (no sources), and the coordinates of the stars. The survey id number is also readout of the pipeline solution file to help cross-validate matching of the disCovErr outputs and the original sources. The emperical gain is read out of the header (for other photometric pipelines which don't perform this estiamte, the gain from DECam is likely sufficient).

Arguments:

base: parent directory and file name prefix for crowdsource results files

date: date_time of the exposure

filt: optical filter used to take the exposure

vers: NOAO community processing version number

ccd: which ccd we are pulling the image for

disCovErr.read_decam — Method

read_decam(base,date,filt,vers,ccd) -> ref_im, w_im, d_im

Read in raw image files associated with exposures obtain on the DarkEnergyCamera. Returns the image, a weighting image, and a quality flag mask image. See NOAO handbook for more details on what is contained in each file and how they are obtained.

Arguments:

base: parent directory and file name prefix for exposure files

date: date_time of the exposure

filt: optical filter used to take the exposure

vers: NOAO community processing version number

ccd: which ccd we are pulling the image for

Example

ref_im, w_im, d_im = read_decam("/n/fink2/decaps/c4d_","170420_040428","g","v1","N14")