disCovErr.jl Documentation

disCovErr.jl Documentation

disCovErr.cov_construct — Method

cov_construct(img, cx, cy; Np::Int=33, widx::Int=129, widy::Int=129) -> cov, μ

Construct a local covariance matrix estimate from img centered around pixel cx and cy. The covariance matrix will be for a square subimage of size Np by Np pixels, yielding cov of size Np^2. The covariance matrix is estimated by taking samples over pixels within a box of size widx and widy centered on the pixel of interest.

If cx and cy are arrays, the first index of the returned cov will be the index of the center point and μ will be the same length as cx.

In the interest of speed for the case of wanting to construct the covariance matrix at many posiitons, the moving mean of the full input image (times shifts of the input image) is precomputed for every entry of the covariance matrix. This may not be efficient for few cx.

Note that the values of widx and widy determine how local an estimate the covariance matrix returned is, but for too small of values of widx and widy, the covariance matrix estimate may be singular (one should always keep Np^2 < widx*widy).

Arguments:

img: Input image for which we desire the covariance matrix estimate

cx: x position (or positions) at which we want a covariance matrix estimate

cy: y position (or positions) at which we want a covariance matrix estimate

Np: size of the image stamp for which the covariance matrix will be estimated

widx: size of sampled region in x used to estimate entries in the covariance matrix

widy: size of sampled region in y used to estimate the entries in the covariance matrix

disCovErr.per_star — Method

per_star(ind) -> per_star_stats()

TBD

disCovErr.per_star_stats — Method

per_star_stats(cov_loc,cxx,cyy,μ,ind) -> per_star_stats()

A bunch o stats