cloudCovErr.jl Documentation

cloudCovErr.jl Documentation

cloudCovErr.add_sky_noise! — Method

add_sky_noise!(testim2,maskim0,skyim3,gain;seed=2021)

Adds noise to the infill that matches the Poisson noise of a rough estimate for the sky background. A random seed to set a local random generator is provided for reproducible unit testing.

Arguments:

testim2: input image which had infilling

maskim0: mask of pixels which were infilled

skyim3: rough estimate of sky background counts

gain: gain of detector to convert from photon count noise to detector noise

seed: random seed for random generator

cloudCovErr.boxsmooth! — Method

boxsmooth!(out, arr, widx::Int, widy::Int, sx::Int, sy::Int)

Boxcar smooths an input image (or paddedview) arr with window size widx by widy. We pass the original image size sx and sy to help handle image views.

Arguments:

out: preallocated output array for the boxcar smoothed image

arr: input array for which boxcar smoothing is computed (generally paddedview)

widx::Int: size of boxcar smoothing window in x

widy::Int: size of boxcar smoothing window in y

sx::Int: x size of the original (unpadded) image

sy::Int: y size of the original (unpadded) image

cloudCovErr.condCovEst_wdiag — Method

condCovEst_wdiag(cov_loc,μ,k,kstar,kpsf2d,data_in,data_w,stars_in) -> [std_w std_wdiag var_wdb resid_mean pred_mean chi20]

Using a local covariance matrix estimate cov_loc and a set of known pixels k and unknown pixels kstar, this function computes a prediction for the mean value of the kstar pixels and the covariance matrix of the kstar pixels. In terms of statistics use to adjust the photometry of a star, we are only interested in the pixels masked as a result of the star (i.e. not a detector defect or cosmic ray nearby) which is kpsf2d. The residual image data_in, the weight image data_w, and a model of the counts above the background coming from the star stars_in for the local patch are also inputs of the function. Correction factors for the photometric flux and flux uncertainities are outputs as well as the chi2 value for the predicted pixels.

Arguments:

cov_loc: local covariance matrix

μ: vector containing mean value for each pixel in the patch

k: unmasked pixels

kstar: masked pixels

kpsf2d: pixels masked due to the star of interest

data_in: residual image in local patch

data_w: weight image in local patch

stars_in: image of counts from star alone in local patch

cloudCovErr.gen_mask_staticPSF! — Method

gen_mask_staticPSF!(maskd, psfstamp, x_stars, y_stars, flux_stars, thr=20)

Generate a mask for an input image (which is usually an image of model residuals) that excludes the cores of stars (which are often mismodeled). In this function, we use a fixed PSF psfstamp for all sources, and adjust the masking fraction based on the stellar flux and a threshold thr. A more general position dependent PSF model could be used with a slight generalization of this function, but is likely overkill for the problem of making a mask.

Arguments:

maskd: bool image to which mask will be added (bitwise or)

psfstamp: simple 2D array of a single PSF to be used for the whole image

x_stars: list of source x positions

y_stars: list of source y positions

flux_stars: list of source fluxes

thr: threshold used for flux-dependent masking

cloudCovErr.prelim_infill! — Method

prelim_infill!(testim,maskim,bimage,bimageI,testim2, maskim2, goodpix; widx = 19, widy=19)

This intial infill replaces masked pixels with a guess based on a smoothed boxcar. For large masked regions, the smoothing scale is increased. If this iteration takes too long/requires too strong of masking, the masked pixels are replaced with the median of the image.

We use 3 copies of the input image and mask image. The first is an untouched view (with reflective boundary condition padding), the second is allocated to hold various smoothings of the image, and the third holds the output image which contains our best infill guess. A final bool array of size corresponding to the image is used to keep track of pixels that have safe infill values.

Arguments:

testim: input image which requires infilling

bimage: preallocated array for smoothed version of input image

testim2: inplace modified ouptut array for infilled version of image input

maskim: input mask indicating which pixels require infilling

bimageI: preallocated array for smoothed mask counting the samples for each estimate

maskim2: inplace modified mask to keep track of which pixels still need infilling

widx::Int: size of boxcar smoothing window in x

widy::Int: size of boxcar smoothing window in y

cloudCovErr.stamp_cutter — Method

stamp_cutter(cxx,cyy,residimIn,w_im,mod_im,skyim,maskim;Np=33) -> data_in, data_w, stars_in, kmasked2d

Cuts out local stamps around each star of the various input images to be used for per star statistics calculations.

Arguments:

cxx: center coorindate x of the stamp

cyy: center coorindate y of the stamp

residimIn: residual image with infilling from which covariance was estimated

w_im: input weight image

mod_im: input model image

skyim: input image of sky background

maskim: input image of masked pixels

Np: size of covariance matrix footprint around each star